+++ samples/004-light-support/src/game.cpp
@@ -65,7 +65,7 @@ namespace light {
 		    std::make_pair<std::string, uint32_t>("./content/textures/light_test/planks.png", 0),
 		    std::make_pair<std::string, uint32_t>("./content/textures/light_test/planksSpec.png", 0)};
 
-		this->_loadingFiles = initialContentFiles.size();
+		this->_loadingFiles = static_cast<int>(initialContentFiles.size());
 		for (auto& f : initialContentFiles) {
 			rawrbox::RESOURCES::loadFileAsync(f.first, f.second, [this]() {
 				this->_loadingFiles--;
